+++ include/dcc/tx/crtp_base.hpp
@@ -18,26 +18,37 @@
 #include "command_station.hpp"
 #include "config.hpp"
 #include "timings.hpp"
-#include "track_outputs.hpp"
+#include "timings_adapter.hpp"
 
 namespace dcc::tx {
 
 /// CRTP base for transmitting DCC
 ///
 /// \tparam T Type to downcast to
-template<typename T>
+/// \tparam D Deque value type
+template<typename T, typename D = Packet>
+requires(std::same_as<D, Packet> || std::same_as<D, Timings>)
 struct CrtpBase {
   friend T;
+
+  using value_type =
+    std::conditional_t<std::same_as<D, Packet>, TimingsAdapter, Timings>;
 
   /// Initialize
   ///
   /// \param  cfg Configuration
-  void init(Config cfg) {
+  void init(Config cfg = {}) {
     assert(cfg.num_preamble >= DCC_TX_MIN_PREAMBLE_BITS &&                 //
            cfg.num_preamble <= DCC_TX_MAX_PREAMBLE_BITS &&                 //
            cfg.bit1_duration >= Bit1Min && cfg.bit1_duration <= Bit1Max && //
            cfg.bit0_duration >= Bit0Min && cfg.bit0_duration <= Bit0Max);  //
     _cfg = cfg;
+    if constexpr (std::same_as<D, Packet>)
+      _idle_packet = TimingsAdapter{make_idle_packet(), _cfg};
+    else if constexpr (std::same_as<D, Timings>)
+      _idle_packet = bytes2timings(make_idle_packet(), _cfg);
+    _first = begin(_idle_packet);
+    _last = cend(_idle_packet);
   }
 
   /// Transmit packet
@@ -57,7 +68,7 @@ struct CrtpBase {
   bool bytes(std::span<uint8_t const> bytes) {
     if (full(_deque)) return false;
     assert(std::size(bytes) <= DCC_MAX_PACKET_SIZE);
-    _deque.push_back(bytes2timings(bytes, _cfg));
+    pushBack(bytes);
     return true;
   }
 
@@ -65,24 +76,33 @@ struct CrtpBase {
   ///
   /// \return Bit duration in µs
   Timings::value_type transmit() {
-    // As long as there are packet timings
-    if (_packet_count < _packet->size()) return packetTiming();
-    // or BiDi timings
-    else if (_cfg.flags.bidi && _bidi_count <= 4uz) return biDiTiming();
+    // Packet timings
+    if (_first != _last) return packetTiming();
+    // Packet end
+    else if constexpr (requires(T t) {
+                         { t.packetEnd() };
+                       })
+      if (!_bidi_count) impl().packetEnd();
 
-    // Deque is empty, send idle packet
-    if (empty(_deque)) _packet = &_idle_packet;
-    // Deque contains packet, send it
+    // BiDi timings
+    if (_cfg.flags.bidi && _bidi_count <= 4uz) return biDiTiming();
+    else _bidi_count = 0uz;
+
+    // Deque is empty, transmit idle packet
+    if (empty(_deque)) {
+      _first = begin(_idle_packet);
+      _last = cend(_idle_packet);
+    }
+    // Deque contains packet, transmit it
     else {
-      _packet = &_deque.front();
+      _first = begin(_deque.front());
+      _last = cend(_deque.front());
       /// \warning
       /// Careful! This only works because of the design of ztl::inplace_deque.
-      /// The slot _packet currently points to will stay valid until the next
-      /// call of pop_front().
+      /// The element currently pointed to will stay valid until the next call
+      /// of pop_front().
       _deque.pop_front();
     }
-    _packet_count = _bidi_count = 0uz;
-
     return packetTiming();
   }
 
@@ -98,22 +118,17 @@ struct CrtpBase {
 
 private:
   constexpr CrtpBase() = default;
-  CommandStation auto& impl() { return static_cast<T&>(*this); }
-  CommandStation auto const& impl() const {
-    return static_cast<T const&>(*this);
-  }
+  auto& impl() { return static_cast<T&>(*this); }
+  auto const& impl() const { return static_cast<T const&>(*this); }
 
   /// Packet timing
   ///
   /// \return Next timings from current packet
   Timings::value_type packetTiming() {
-    if constexpr (TrackOutputs<T>)
-      !(_packet_count % 2uz)
-        ? impl().trackOutputs(false ^ _cfg.flags.invert, // First half bit
-                              true ^ _cfg.flags.invert)
-        : impl().trackOutputs(true ^ _cfg.flags.invert, // Second half bit
-                              false ^ _cfg.flags.invert);
-    return (*_packet)[static_cast<Timings::size_type>(_packet_count++)];
+    toggleTrackOutputs();
+    auto const retval{*_first};
+    ++_first;
+    return retval;
   }
 
   /// BiDi timing
@@ -123,47 +138,81 @@ private:
     switch (_bidi_count++) {
       // Send half a 1 bit
       case 0uz:
-        if constexpr (TrackOutputs<T>)
-          impl().trackOutputs(false ^ _cfg.flags.invert,
-                              true ^ _cfg.flags.invert);
+        toggleTrackOutputs();
         return static_cast<Timings::value_type>(bidi::Timing::TCS);
 
       // Cutout start
       case 1uz:
-        if constexpr (TrackOutputs<T>)
-          impl().trackOutputs(false ^ _cfg.flags.invert,
-                              false ^ _cfg.flags.invert);
-        impl().biDiStart();
+        toggleTrackOutputs();
+        if constexpr (requires(T t) {
+                        { t.biDiStart() };
+                      })
+          impl().biDiStart();
         return static_cast<Timings::value_type>(bidi::Timing::TTS1 -
                                                 bidi::Timing::TCS);
 
       // Channel1 start
       case 2uz:
-        impl().biDiChannel1();
+        if constexpr (requires(T t) {
+                        { t.biDiChannel1() };
+                      })
+          impl().biDiChannel1();
         return static_cast<Timings::value_type>(bidi::Timing::TTS2 -
                                                 bidi::Timing::TTS1);
 
       // Channel2 start
       case 3uz:
-        impl().biDiChannel2();
+        if constexpr (requires(T t) {
+                        { t.biDiChannel2() };
+                      })
+          impl().biDiChannel2();
         return static_cast<Timings::value_type>(bidi::Timing::TTC2 -
                                                 bidi::Timing::TTS2);
 
       // Cutout end
       default:
-        impl().biDiEnd();
+        if constexpr (requires(T t) {
+                        { t.biDiEnd() };
+                      })
+          impl().biDiEnd();
         return static_cast<Timings::value_type>(bidi::Timing::TCE -
                                                 bidi::Timing::TTC2);
     }
   }
 
-  static constexpr Timings _idle_packet{packet2timings(make_idle_packet())};
+  /// Add packet or timings to deque
+  ///
+  /// \param  bytes Bytes containing DCC packet
+  void pushBack(std::span<uint8_t const> bytes) {
+    if constexpr (std::same_as<D, Packet>) _deque.push_back({bytes, _cfg});
+    else if constexpr (std::same_as<D, Timings>)
+      _deque.push_back(bytes2timings(bytes, _cfg));
+  }
 
-  ztl::inplace_deque<Timings, DCC_TX_DEQUE_SIZE> _deque{};
-  Timings const* _packet{&_idle_packet};
-  size_t _packet_count{};
-  size_t _bidi_count{};
-  Config _cfg{};
+  /// Toggle track outputs
+  void toggleTrackOutputs() {
+    if constexpr (requires(T t, bool N, bool P) {
+                    { t.trackOutputs(N, P) };
+                  }) {
+      // By default the phase is "positive", so P > N for the first half bit.
+      impl().trackOutputs(_polarity, !_polarity);
+      _polarity = !_polarity;
+    }
+  }
+
+  /// Deque
+  ztl::inplace_deque<value_type, DCC_TX_DEQUE_SIZE> _deque{};
+
+  /// Idle packet
+  value_type _idle_packet{};
+
+  /// Iterators
+  decltype(std::begin(_idle_packet)) _first{std::begin(_idle_packet)};
+  decltype(std::cend(_idle_packet)) _last{std::cend(_idle_packet)};
+
+  size_t _bidi_count{}; ///< Count BiDi timings
+  Config _cfg{};        ///< Configuration
+  bool _polarity{};     ///< Track polarity
 };
 
 } // namespace dcc::tx