+++ tests/tx/capacity.cpp
@@ -1,3 +1,6 @@
 #include "tx_test.hpp"
 
-TEST_F(TxTest, capacity) { EXPECT_EQ(_mock.capacity(), DCC_TX_DEQUE_SIZE); }
+TEST_F(TxTest, capacity) {
+  EXPECT_EQ(_packet_mock.capacity(), DCC_TX_DEQUE_SIZE);
+  EXPECT_EQ(_timings_mock.capacity(), DCC_TX_DEQUE_SIZE);
+}