+++ tests/tx/packet.cpp
@@ -1,6 +1,7 @@
 #include "tx_test.hpp"
 
 TEST_F(TxTest, packet) {
-  for (auto i{DCC_TX_DEQUE_SIZE}; i-- > 0uz;) EXPECT_TRUE(_mock.packet({}));
-  EXPECT_FALSE(_mock.packet({}));
+  auto packet{dcc::make_idle_packet()};
+  EXPECT_TRUE(_packet_mock.packet(packet));
+  EXPECT_TRUE(_timings_mock.packet(packet));
 }