+++ tests/tx/tx_test.hpp
@@ -11,5 +11,9 @@ protected:
   TxTest();
   virtual ~TxTest();
 
-  NiceMock<TxMock> _mock;
+  dcc::tx::Config _cfg{.num_preamble = DCC_TX_MIN_PREAMBLE_BITS + 1u,
+                       .bit1_duration = 59u,
+                       .bit0_duration = 113u};
+  NiceMock<TxMock<dcc::Packet>> _packet_mock;
+  NiceMock<TxMock<dcc::tx::Timings>> _timings_mock;
 };