+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+##
+- `tx::CrtpBase` stores packets as either bytes or timings ([#57](https://github.com/ZIMO-Elektronik/DCC/issues/57))
+
 ## 0.41.1
 - Update to ZTL 0.21.0